remise en forme ecran. introduction note archers. nouvelle table note_archer

+++ app/src/main/java/com/herault/comptecible/utils/Db_resultat.java
@@ -4,7 +4,6 @@ import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.provider.BaseColumns;
-import android.util.Log;
 
 class Db_resultat extends SQLiteOpenHelper {
 
@@ -13,12 +12,29 @@ class Db_resultat extends SQLiteOpenHelper {
      * The static string to create the database.
      */
     private static final String DATABASE_CREATE_TABLE_ARCHERS = "create table "
-            + Constants.ARCHERS + "(" + Constants.KEY_ID_COLUMN
-            + " INTEGER primary key autoincrement, " + Constants.KEY_COL_NAME + " TEXT UNIQUE)  ";
+            + Constants.ARCHERS + "(" + Constants.KEY_ID_ARCHERS
+            + " INTEGER primary key autoincrement, "
+            + Constants.KEY_COL_NAME + " TEXT UNIQUE, "
+            + Constants.KEY_COL_INFORMATION + " TEXT " + ")  ";
 
     private static final String DATABASE_CREATE_TABLE_ARCHERS_ROUND = "create table "
-            + Constants.ARCHERS_ROUND + "(" + Constants.KEY_ID_COLUMN
-            + " INTEGER primary key autoincrement, " + Constants.KEY_COL_NAME + " TEXT UNIQUE)  ";
+            + Constants.ARCHERS_ROUND + "(" + Constants.KEY_ID_ARCHERS
+            + " INTEGER primary key autoincrement, "
+            + Constants.KEY_COL_NAME + " TEXT UNIQUE, "
+            + Constants.KEY_COL_INFORMATION + " TEXT " + ")  ";
+
+
+
+    private static final String DATABASE_CREATE_TABLE_ARCHER_NOTES = "create table "
+            + Constants.ARCHERS_NOTE + "(" + Constants.KEY_ID_NOTE
+            + " INTEGER primary key autoincrement, "
+            + Constants.KEY_COL_ID_ARCHERS + " INTEGER, "
+            + Constants.KEY_COL_NAME_NOTE + " TEXT, "
+            + " CONSTRAINT c "
+            + " FOREIGN KEY ( " + Constants.KEY_COL_ID_ARCHERS + " ) REFERENCES " + Constants.ARCHERS
+            + " ( " + Constants.KEY_ID_ARCHERS + " ) ON DELETE CASCADE"+ ")  "
+            ;
+
 
     private static final String DATABASE_CREATE_TABLE_ROUNDS = "create table "
             + Constants.ROUNDS + "(" + Constants.KEY_ID_ROUNDS
@@ -40,10 +56,10 @@ class Db_resultat extends SQLiteOpenHelper {
             + " ( " + Constants.KEY_ID_ROUNDS + " ) ON DELETE CASCADE,"
             + "CONSTRAINT b"
             + " FOREIGN KEY ( " + Constants.KEY_COL_ID_NAME + " ) REFERENCES " + Constants.ARCHERS
-            + " ( " + Constants.KEY_ID_COLUMN + " ) ON DELETE CASCADE)";
+            + " ( " + Constants.KEY_ID_ARCHERS + " ) ON DELETE CASCADE)";
 
     private static final String DATABASE_CREATE_TABLE_COMPTE_CIBLE_ = "create table "
-            + Constants.COMPTE_CIBLE + "(" + Constants.KEY_ID_COLUMN
+            + Constants.COMPTE_CIBLE + "(" + Constants.KEY_ID_ARCHERS
             + " INTEGER primary key autoincrement, " + Constants.KEY_COL_COMPTE_NAME + " TEXT UNIQUE,"
             + Constants.KEY_COL_COMPTE_VALUE + " TEXT)  ";
 
@@ -61,8 +77,10 @@ class Db_resultat extends SQLiteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase db) {
         // Create the new database using the SQL string Database_create
-        db.execSQL(DATABASE_CREATE_TABLE_ARCHERS_ROUND);
+
         db.execSQL(DATABASE_CREATE_TABLE_ARCHERS);
+        db.execSQL(DATABASE_CREATE_TABLE_ARCHER_NOTES);
+        db.execSQL(DATABASE_CREATE_TABLE_ARCHERS_ROUND);
         db.execSQL(DATABASE_CREATE_TABLE_COMPTE_CIBLE_);
         db.execSQL(DATABASE_CREATE_TABLE_ROUNDS);
         db.execSQL(DATABASE_CREATE_TABLE_RESULTATS);
@@ -73,14 +91,24 @@ class Db_resultat extends SQLiteOpenHelper {
      /*  Log.w("DBOpenHelper", "Mise à jour de la version " + oldVersion
                 + " vers la version " + newVersion
                 + ", les anciennes données seront détruites "); */
-        // Drop the old database
-        db.execSQL("DROP TABLE IF EXISTS " + Constants.ARCHERS);
-        db.execSQL("DROP TABLE IF EXISTS " + Constants.ARCHERS_ROUND);
-        db.execSQL("DROP TABLE IF EXISTS " + Constants.RESULTATS);
-        db.execSQL("DROP TABLE IF EXISTS " + Constants.COMPTE_CIBLE);
-        db.execSQL("DROP TABLE IF EXISTS " + Constants.ROUNDS);
-        // Create the new one
-        onCreate(db);
+        if (oldVersion == 23 && newVersion == 24)
+            {
+            db.execSQL("ALTER TABLE "+ Constants.ARCHERS + " ADD "  + Constants.KEY_COL_INFORMATION + " TEXT"  );
+            db.execSQL("ALTER TABLE "+ Constants.ARCHERS_ROUND + " ADD "  + Constants.KEY_COL_INFORMATION + " TEXT"  );
+            db.execSQL(DATABASE_CREATE_TABLE_ARCHER_NOTES);
+            }
+        else
+            {
+            // Drop the old database
+            db.execSQL("DROP TABLE IF EXISTS " + Constants.ARCHERS);
+            db.execSQL("DROP TABLE IF EXISTS " + Constants.ARCHERS_NOTE);
+            db.execSQL("DROP TABLE IF EXISTS " + Constants.ARCHERS_ROUND);
+            db.execSQL("DROP TABLE IF EXISTS " + Constants.RESULTATS);
+            db.execSQL("DROP TABLE IF EXISTS " + Constants.COMPTE_CIBLE);
+            db.execSQL("DROP TABLE IF EXISTS " + Constants.ROUNDS);
+            // Create the new one
+            onCreate(db);
+        }
         // or do a smartest stuff
     }
 
@@ -92,33 +120,33 @@ class Db_resultat extends SQLiteOpenHelper {
         // The database name
         public static final String DATABASE_NAME = "resultat.db";
         // The database version
-        public static final int DATABASE_VERSION = 23;
-
+        public static final int DATABASE_VERSION = 24;
+// -----------------------
         // The table Name
         public static final String ARCHERS = "archers";
+
         // ## Column name ##
         // My Column ID and the associated explanation for end-users
-        public static final String KEY_ID_COLUMN = "_id";// Mandatory
+        public static final String KEY_ID_ARCHERS = "_id";// Mandatory
         // My Column Name and the associated explanation for end-users
         public static final String KEY_COL_NAME = "name";
-        // Indexes des colonnes
-        // The index of the column ID
-        public static final int ID_COLUMN = 1;
-        // The index of the column NAME
-        public static final int COL_NAME = 2;
+        public static final  String KEY_COL_INFORMATION = "information" ;
 
+        // -----------------------
+        // The table Name
+        public static final String ARCHERS_NOTE = "archers_notes";
+
+        // ## Column name ##
+        // My Column ID and the associated explanation for end-users
+        public static final String KEY_ID_NOTE = "_id";// Mandatory
+        public static final  String KEY_COL_ID_ARCHERS = "_id_archers" ;
+        // My Column Name and the associated explanation for end-users
+        public static final String KEY_COL_NAME_NOTE = "note";
+
+
+// -----------------------
         // The table Name
         public static final String ARCHERS_ROUND = "archers_round";
-        // ## Column name ##
-        // My Column ID and the associated explanation for end-users
- /*       public static final String KEY_ID_COLUMN = "_id";// Mandatory
-        // My Column Name and the associated explanation for end-users
-        public static final String KEY_COL_NAME = "name";
-        // Indexes des colonnes
-        // The index of the column ID
-        public static final int ID_COLUMN = 1;
-        // The index of the column NAME
-        public static final int COL_NAME = 2; same as table ARCHER*/
 
         // The table Name
         public static final String ROUNDS = "rounds";
@@ -129,15 +157,8 @@ class Db_resultat extends SQLiteOpenHelper {
         public static final String KEY_COL_ROUND_NAME = "name_round";
 
         public static final String KEY_COL_ROUND_TYPE = "type_round";
-        // Indexes des colonnes
-        // The index of the column ID
-        public static final int COL_ROUNDS_ID_ = 1;
-        // The index of the column NAME
-        public static final int COL_ROUND_NAME = 2;
 
-
-
-
+// -----------------------
         // All single value for application like Number of arrow by round
         // The table Name
         public static final String COMPTE_CIBLE = "compteCible";
@@ -149,15 +170,7 @@ class Db_resultat extends SQLiteOpenHelper {
         // Type of round
         public static final String KEY_COL_COMPTE_VALUE = "value";
 
-        // Indexes des colonnes
-        // The index of the column ID
-        // public static final int ID_COLUMN = 1;
-        // The index of the column NAME
-        public static final int COL_NAME_COMPTE = 2;
-
-        public static final int COL_VALUE_COMPTE = 3;
-
-
+ // -----------------------
         // The table Name
         public static final String RESULTATS = "resultats";
         // ## Column name ##
@@ -176,21 +189,8 @@ class Db_resultat extends SQLiteOpenHelper {
         // My Column id_Name
         public static final String KEY_COL_ID_NAME = "id_name";
 
-        // Indexes des colonnes
-        // The index of the column ID
-        public static final int COL_ID1 = 1;
-        // The index of the column ROUND
-        public static final int COL_ROUND = 2;
-        // The index of the column ARROW
-        public static final int COL_ARROW = 3;
-        // The index of the column VALUE
-        public static final int COL_VALUE = 4;
-        // The index of the column ARROW
-        public static final int COL_X = 5;
-        // The index of the column VALUE
-        public static final int COL_Y = 6;
-        // The index of the column NAME
-        public static final int COL_ID_NAME = 7;
+ // -----------------------
+
 
 
     }
